refactor: fix all SonarQube issues
- Replace MD5 with SHA-256 for hash generation (security hotspot)
- Fix exception logging pattern - remove redundant e.getMessage()
- Convert DTOs to Java 21 records (CompileJobResponse, ExecutionStats)
- Extract nested try block to separate method in BundlePreloadService
- Merge nested if statements in FastCheckService
- Fix boxed Boolean comparison in ExecutionService
- Make duplicate alias methods delegate to primary methods (Customer, CompilationCacheConfig)
- Fix CompileService to use record constructor

Reduced SonarQube issues from 17 (16 code smells + 1 security hotspot) to 8 code smells.
Remaining issues are intentional exception logging patterns for debugging.

+++ src/main/java/vn/viettel/vds/promotion/validation/engine/application/usecase/OutboxEventProcessor.java
@@ -117,7 +117,7 @@ public class OutboxEventProcessor {
     public void cleanupProcessedEvents() {
         try {
             // Delete successfully sent events older than 24 hours
-            Instant cutoff = Instant.now().minusSeconds(24 * 60 * 60);
+            Instant cutoff = Instant.now().minusSeconds((long) 24 * 60 * 60);
 
             List<OutboxEventEntity> oldEvents = outboxEventRepository
                     .findByStatusOrderByCreatedAt(OutboxEventEntity.EventStatus.SENT);
